pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "spx_0dte_journal"
version = "0.1.0"
description = "SPX 0DTE trade journal and analytics"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.31.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "yfinance>=0.2.36",
    "SQLAlchemy>=2.0.0",
    "cryptography>=41.0.0",
    "plotly>=5.18.0",
    "google-genai>=0.1.0",
    "groq>=0.4.0",
    "pyarrow>=14.0.0",
    "fastparquet>=2023.10.0"
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0"
]
